Expression is now a statement
An expression followed by a semicolon is now considered a statement.

Furthermore assignment is now considered an expression. This allows an
assigment to appear inside conditions and inside another assignment's
expression.

This changes in the grammar allow binary expressions (such as 'x + 3') and
unary expressions (such as !x) be statements on their own. This is considered
an error, but it will be caught only in future semantic analysis passes.