Optimization to lexer
Now the use can specify whether or not to save the lexeme of the source
file into the token (this permits some memory optimization).

The token types strings and saved lexeme now get interned via
String#intern method. This allows space optimization and comparison via
== and != instead of equals (ths works because interned strings are
mapped to the same memory object) increasing the performance of the
parser.